Fixes
Fixed a bug which causes undefined behavior if the user starts creating or deleting files in the Logs folder. 

Added: A check to make sure we only save files that contain the text Alert Log
and the last 3 characters are "log"

Added : Increase CURL pause to 4 seconds. Necessary because some photos take longer to save depending on the amount of info on the screen. The SC screenshot function might only save the file after the study function returns.