Add files via upload
Resolved an issue regarding the use of a particular setting that changes the location of the chart number in the text added to the alert log file. 

This setting is Global Settings > General Settings > Charts > Show Chart Number First on Chart Name

It should now be fine to either enable or disable this setting.

One other note to add is that on Windows. If you ever receive an error in the message log about saving the image file. Carefully inspect the text of the image file that it tried to save. It is likely that an invalid character was detected in the file name. If you are able to reproduce an issue like this please let us know at support@verrillotrading.com.