Add files via upload
This Advanced Custom Study was written by Christian at VerrilloTrading during December of 2024.

Study Description:

This study changes the symbol of the current chart to the symbol of the target chart that is set in the input setting of the study.

The purpose of this study is to solve an issue when users configure historical and intraday charts within the same chartbook and they want to use chart linking to keep the historical charts linked, and the intraday chrats linked.

If a user runs that type of configuration, it becomes impossible to use chart linking to automaticlly change the symbol of the intraday chart if the historical chart symbol is changed.

This study works the same as if the two charts were part of the same chart linking group with the symbol linked. Except that it bypasses chart linking and handles the symbol switching on it's own.

----
First it is important to understand how the study works one way. Then Bidirectional Symbol Switching will be explained.
Simple Unidirectional Symbol Switching:

1. The Sierra Chart chartbook contains two charts, Chart #1 and Chart #2.
2. This study is placed on Chart #1.
3. The input setting Chart Number To Monitor Symbol Change is set to target Chart #2.

4. This means that when the symbol on Chart #2 is changed to a new symbol, the symbol on Chart #1 will change to match the symbol on Chart #2.
This is fine except that if the user decides to change the symbol on Chart #1 instead. The symbol of Chart #2 will not automatically change to match the symbol on Chart #1 unless we use the Bidirectional method.

Bidirectional Chart Switching Configuration:

Bidirectional Symbol Switching refers to when the symbol is changed on either Chart #1 or Chart #2. The other chart also changes it's symbol.
Bidirectional Symbol Switching is supported here if this study is placed on both charts that you want to keep the symbols linked. The charts must be set to target eachother using the input setting Chart Number To Monitor Symbol Change.
1. This functionality is enabled by default. All that is required is that the study is placed on both charts that you want to automatically switch when one chart or the other is changed.

2. In order to do this, the study needs to be placed on both Chart #1 and Chart #2. The input setting Chart Number To Monitor Symbol Change for Chart #1 will be set a target chart #2. For Chart #2 it should be set to target Chart #1.

This Advanced Custom Study is open source and free to use.

For any inquiries or issues please send them to support@verrillotrading.com.

Kind Regards,

-Christian

VerrilloTrading, Content Creator, Programmer