Fixed a non responsive study input
The main patch is was in the previous commit. Note that the commits that get uploaded to github are big commits and my system has the larger list of commits during development phase. so if you are trying to see what changes were made using just the commits posted to github, good luck! 

Notes from previous commit: 

-----------------------                                                                                                         DONE: Add Support for draw the levels realtime when new levels are created                                                      intraday.                                                                                                                                                                                                                                                       DONE: Line 6556 test using multiple timezone configs                                                                                                                                                                                                            
DONE: FIX HISTORICAL CODE TO MAKE SURE WE REMEMBER MOST RECENT DAY LINE NUMBERS                                                 
And also handle different timezones where the time in seconds for the level                                                     
might be 0 in that case it is gracefully handled.                                                                               
                                                                                                                                
DONE: See if we can draw globex close time in realtime on some symbols like stocks after hours                                  
                                                                                                                                
CANNOT DRAW GLOBEX CLOSE IN REALTIME ON GLOBEX CLOSE SYMBOLS. Just recalculate                                                  
the chart it and it should work.                                                                                                
                                                                                                                                
DONE: Add input to draw Levels over the weekend or not. Enable this for symbols                                                 
that continue trading over the weekend.                                                                                         
                                                                                                                                
Determined that the study is not able to draw the levels for futures contracts that have expired a long time ago.