Add files via upload
Added a fix for a new setting in Sierra Chart under Global Settings > General Settings > Alerts > User Defined Name > Use User Defined Name Only in Alert Message. 

This setting changes the format in the alert log files when alerts are triggered. 

Right now the fix for this is temporary. We need to communicate a few inconsistencies regarding alert log files to SC before writing a more concrete fix for support with that menu item. 

For the moment it is recommended to leave option set to: No 

Do consider getting added to our list of authorized study users by sending an email to support@verrillotrading.com